refactor: Move probe_fudge into fudge.rs
